CambioTipoDato_PesablesDetalle_Columna_Diametro

+++ Aponus Web API/Mapping/EspecificacionesComponentes.cs
@@ -23,7 +23,7 @@ namespace Aponus_Web_API.Mapping
         public int? Perfil { get; set; }
 
         [JsonProperty(PropertyName = "diametro" , NullValueHandling = NullValueHandling.Ignore)]
-        public decimal? Diametro { get; set; }
+        public int? Diametro { get; set; }
 
         [JsonProperty(PropertyName = "Altura" , NullValueHandling = NullValueHandling.Ignore)]
         public decimal? Altura { get; set; }
